Corrected gitignore problem + Removed github action + Added tokenize.py for good (#6)
## 🎯 Goal / Problem

Corrected gitignore problem + Removed github action + Added tokenize.py
for good

## 💡 Proposed solution

Corrected gitignore problem + Removed github action + Added tokenize.py
for good

## 🔬 How to test

<!-- Tell the reviewer how you tested that the change works as expected,
and/or the steps they can follow to test it themselves. -->

## 📚 References

<!-- Provide links to resources that were useful to address this
problem/goal, and/or the solution proposed in this PR, if any. -->

## ℹ️ Notes for reviewer

<!-- If you expect the reviewer to be surprised about something (e.g.
the size of the diff, a dirty workaround...), here is a good place to
mention it and provide a rationale. You can also use this section to
suggest the most efficient way to review this PR, e.g. commit by commit,
or starting by a particular file. -->

## ✅ Merge checklist

I/we self-reviewed this PR so that:

- [ ] the **title** is a [conventional
commit](https://www.conventionalcommits.org/) and will still make sense
for the whole team, a month from now;
- [ ] the **description** provides useful context for an effective
review;
- [ ] the **code diff** is cleaned up, and only addresses the goal and
intended scope of this PR;